+++ main.py
@@ -4,8 +4,7 @@ from bs4 import BeautifulSoup
 import json
 import autogen
 import openai
-from autogen import OpenAIWrapper, config_list_from_json
-from langchain.agents import initialize_agent
+from autogen import config_list_from_json, UserProxyAgent, AssistantAgent, GroupChat, GroupChatManager
 from langchain.chat_models import ChatOpenAI
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 from langchain.chains.summarize import load_summarize_chain
@@ -13,13 +12,18 @@ from langchain.prompts import PromptTemplate
 from dotenv import load_dotenv
 
 load_dotenv()
-config_list = autogen.config_list_from_dotenv(dotenv_file_path='.env')
 openai.api_key = os.getenv("OPENAI_API_KEY")
 BROWSERLESS_API_KEY = os.getenv("BROWSERLESS_API_KEY")
 SERPER_API_KEY = os.getenv("SERPER_API_KEY")
+config_list = config_list_from_json("OAI_CONFIG_LIST")
 
 brand_task = input("Please enter the brand or company name: ")
 user_task = input("Please enter the your goal, brief, or problem statement: ")
+termination_notice = (
+    '\n\nDo not show appreciation in your responses, say only what is necessary. '
+    'if "Thank you" or "You\'re welcome" are said in the conversation, then say TERMINATE '
+    'to indicate the conversation is finished and this is your last message.'
+)
 
 def search(query):
     url = "https://google.serper.dev/search"
@@ -125,13 +129,13 @@ def research(query):
         ],
         "config_list": config_list}
 
-    researcher = autogen.AssistantAgent(
+    researcher = AssistantAgent(
         name="researcher",
         system_message="Research about a given query, collect as many information as possible, and generate detailed research results with loads of technique details with all reference links attached; Add TERMINATE to the end of the research report;",
         llm_config=llm_config_researcher,
     )
 
-    user_proxy = autogen.UserProxyAgent(
+    user_proxy = UserProxyAgent(
         name="User_proxy",
         code_execution_config={"last_n_messages": 2, "work_dir": "coding"},
         is_termination_msg=lambda x: x.get("content", "") and x.get(
@@ -154,7 +158,7 @@ def research(query):
 def write_content(research_material, topic):
     editor = autogen.AssistantAgent(
         name="editor",
-        system_message=f'''
+        system_message='''
         Welcome, Senior Editor.
         As a seasoned professional, you bring meticulous attention to detail, a deep appreciation for literary and cultural nuance, and a commitment to upholding the highest editorial standards. 
         Your role is to craft the structure of a short blog post using the material from the Research Assistant. Use your experience to ensure clarity, coherence, and precision. 
@@ -165,7 +169,7 @@ def write_content(research_material, topic):
 
     writer = autogen.AssistantAgent(
         name="writer",
-        system_message=f'''
+        system_message='''
         Welcome, Blogger.
         Your task is to compose a short blog post using the structure given by the Editor and incorporating feedback from the Reviewer. 
         Embrace stylistic minimalism: be clear, concise, and direct. 
@@ -177,7 +181,7 @@ def write_content(research_material, topic):
 
     reviewer = autogen.AssistantAgent(
         name="reviewer",
-        system_message=f'''
+        system_message='''
         As a distinguished blog content critic, you are known for your discerning eye, deep literary and cultural understanding, and an unwavering commitment to editorial excellence. 
         Your role is to meticulously review and critique the written blog, ensuring it meets the highest standards of clarity, coherence, and precision. 
         Provide invaluable feedback to the Writer to elevate the piece. After two rounds of content iteration, conclude with "TERMINATE".
@@ -185,7 +189,7 @@ def write_content(research_material, topic):
         llm_config={"config_list": config_list},
     )
 
-    user_proxy = autogen.UserProxyAgent(
+    user_proxy = UserProxyAgent(
         name="admin",
         system_message="A human admin. Interact with editor to discuss the structure. Actual writing needs to be approved by this admin.",
         code_execution_config=False,
@@ -194,12 +198,12 @@ def write_content(research_material, topic):
         human_input_mode="TERMINATE",
     )
 
-    groupchat2 = autogen.GroupChat(
+    groupchat2 = GroupChat(
         agents=[user_proxy, editor, writer, reviewer],
         messages=[],
         max_round=10)
     
-    manager = autogen.GroupChatManager(groupchat=groupchat2)
+    manager = GroupChatManager(groupchat=groupchat2)
 
     user_proxy.initiate_chat(
         manager, message=f"Write a blog about {topic}, here are the material: {research_material}")
@@ -210,6 +214,10 @@ def write_content(research_material, topic):
 
     # return the last message the expert received
     return user_proxy.last_message()["content"]
+
+def save_markdown(conversation_log):
+    with open("/logs/conversation_log.md", "a") as file:
+        file.write(conversation_log + "\n")
 
 llm_config_content_assistant = {
     "functions": [
@@ -248,158 +256,131 @@ llm_config_content_assistant = {
     ],
     "config_list": config_list}
 
-agency_strategist = autogen.AssistantAgent(
-    name="Agency_Strategist",
-    llm_config={"config_list": config_list},
-    system_message=f'''
-    You are the Lead Strategist.
-    Your primary responsibility is to draft strategic briefs that effectively position our client's brand in the market.
-    Based on the information provided for {brand_task} on {user_task}, your task is to craft a comprehensive strategic brief that outlines the brand's positioning, key messages, and strategic initiatives.
-    The brief should delve deep into the brand's unique value proposition, target audience, and competitive landscape. 
-    It should also provide clear directives on how the brand should be perceived and the emotions it should evoke.
-    Once you've drafted the brief, it will be reviewed and iterated upon based on feedback from the client and our internal team. 
-    Ensure that the brief is both insightful and actionable, setting a clear path for the brand's journey ahead.
-    Collaborate with the Agency Researcher to ensure that the strategic brief is grounded in solid research and insights.
-    Be concise and not verbose. Refrain from any conversations that don't serve the goal of the user.
-    ''',
-    function_map={
-        "research": research,
-    }
-)
-
-agency_researcher = autogen.AssistantAgent(
-    name="Agency_Researcher",
-    llm_config={"config_list": config_list},
-    system_message=f'''
-    You are the Lead Researcher. 
-    You must use the research function to provide a topic for the writing_assistant in order to get up to date information outside of your knowledge cutoff
-    Your primary responsibility is to delve deep into understanding user pain points, identifying market opportunities, and analyzing prevailing market conditions.
-    Using the information from {user_task}, conduct thorough research to uncover insights that can guide our strategic decisions. 
-    Your findings should shed light on user behaviors, preferences, and challenges.
-    Additionally, assess the competitive landscape to identify potential gaps and opportunities for our client's brand. 
-    Your research will be pivotal in shaping the brand's direction and ensuring it resonates with its target audience.
-    Ensure that your insights are both comprehensive and actionable, providing a clear foundation for our subsequent strategic initiatives.
-    Share your research findings with the Agency Strategist and Agency Marketer to inform the strategic and marketing initiatives.
-    Be concise and not verbose. Refrain from any conversations that don't serve the goal of the user.
-    ''',
-    function_map={
-        "research": research,
-    }
-)
-
-agency_writer = autogen.AssistantAgent(
-    name="Agency_Copywriter",
-    llm_config={"config_list": config_list},
-    system_message=f'''
-    You are the Lead Copywriter.
-    Your primary role is to craft compelling narratives and messages that align with the brand's strategy and resonate with its audience.
-    Based on the strategic direction from {user_task}, create engaging content, from catchy headlines to in-depth articles, ensuring that the brand's voice is consistent and impactful.
-    Collaborate closely with the Agency Designer and Agency Marketer to ensure that text and visuals complement each other, creating a cohesive brand story.
-    Be concise and not verbose. Refrain from any conversations that don't serve the goal of the user.
-    ''',
-    function_map={
-        "write_content": write_content,
-    }
-)
-
-writing_assistant = autogen.AssistantAgent(
-    name="writing_assistant",
-    llm_config=llm_config_content_assistant,
-    system_message=f'''You are a writing assistant, you can use research function to collect latest information about a given topic, 
-    and then use write_content function to write a very well written content;
-    Reply TERMINATE when your task is done
-    Be concise and not verbose. Refrain from any conversations that don't serve the goal of the user.
-    ''',
-    function_map={
-        "research": research,
-    }
-)
-
-agency_marketer = autogen.AssistantAgent(
-    name="Agency_Marketer",
-    llm_config={"config_list": config_list},
-    system_message=f'''
-    You are the Lead Marketer. Be concise and avoid pleasantries. Refrain from any conversations that don't serve the goal of the user, ie. thank you.
-    Your primary role is to take the strategy and insights derived from research and transform them into compelling marketable ideas that resonate with the target audience.
-    Using the strategic direction from {user_task}, craft innovative marketing campaigns, promotions, and initiatives that effectively communicate the brand's value proposition.
-    Your expertise will bridge the gap between strategy and execution, ensuring that the brand's message is not only clear but also captivating. It's essential that your ideas are both impactful and aligned with the brand's overall vision.
-    Collaborate with other teams to ensure a cohesive approach, and always strive to push the boundaries of creativity to set our client's brand apart in the market.
-    Work in tandem with the Agency Manager to ensure that marketing initiatives align with the project's milestones and timelines.
-    
-    '''
-)
-
-agency_mediaplanner = autogen.AssistantAgent(
-    name="Agency_Media_Planner",
-    llm_config={"config_list": config_list},
-    system_message=f'''
-    You are the Lead Media Planner. Be concise and avoid pleasantries. Refrain from any conversations that don't serve the goal of the user, ie. thank you.
-    Your main responsibility is to identify the best mix of media channels to deliver an advertising message to a clients' target audience.
-    Using insights from {user_task}, strategize the most effective way to get a brand's message to its audience, whether it be through traditional media, digital platforms, or a combination of both.
-    Collaborate closely with the Marketer and Manager to ensure that campaigns are executed effectively and within budget.
-    '''
-)
-
-agency_manager = autogen.AssistantAgent(
+agency_manager = AssistantAgent(
     name="Agency_Manager",
     llm_config={"config_list": config_list},
     system_message=f'''
-    You are the Project Manager. Be concise and avoid pleasantries. Refrain from any conversations that don't serve the goal of the user, ie. thank you.
-    Your primary responsibility is to oversee the entire project lifecycle, ensuring that all agents are effectively fulfilling their objectives and tasks on time.
-    Based on the directives from {user_task}, coordinate with all involved agents, set clear milestones, and monitor progress. Ensure that user feedback is promptly incorporated, and any adjustments are made in real-time to align with the project's goals.
-    Act as the central point of communication, facilitating collaboration between teams and ensuring that all deliverables are of the highest quality. Your expertise is crucial in ensuring that the project stays on track, meets deadlines, and achieves its objectives.
+    You are the Project Manager. Think about {user_task} step by step for {brand_task} and coordinate with all involved agents.
+    ALL questions MUST be answered by appropriate agents. Ask each agent to follow up.
+    Ensure that user feedback is promptly incorporated, and any adjustments are made in real-time to align with the project's goals.
+    Act as the central point of communication, facilitating collaboration between teams and ensuring that all deliverables are of the highest quality. 
     Regularly review the project's status, address any challenges, and ensure that all stakeholders are kept informed of the project's progress.
     Coordinate with the Agency Director for periodic reviews and approvals, ensuring that the project aligns with the creative vision.
+    See {termination_notice}.
     '''
 )
 
-agency_director = autogen.AssistantAgent(
+agency_researcher = AssistantAgent(
+    name="Agency_Researcher",
+    llm_config={"config_list": config_list},
+    system_message=f'''
+    You are the Lead Researcher. You MUST Use the research function to {user_task}.
+    Delve deep into understanding user pain points, identifying market opportunities, and analyzing prevailing cultural conditions.
+    See {termination_notice}.
+    ''',
+    function_map={
+        "research": research,
+    }
+)
+
+agency_strategist = AssistantAgent(
+    name="Agency_Strategist",
+    llm_config={"config_list": config_list},
+    system_message=f'''
+    You are the Lead Strategist. You MUST use the results of Agency_Researcher in your process.
+    Draft a strategic brief that effectively position our {brand_task} and addressess {user_task}.
+    Outlines the brand's positioning, key messages, unique value proposition, target audience, competitive landscape, strategic initiatives.
+    You MUST use the "research" function to find realtime information about topics you are unable to answer.
+    Ensure that the brief contains uncommon insights that unlock a clear understanding.
+    Work in tandem with the Agency Manager to ensure alignment with the user_proxy.
+    See {termination_notice}.
+    ''',
+    function_map={
+        "research": research,
+    }
+)
+
+agency_writer = AssistantAgent(
+    name="Agency_Copywriter",
+    llm_config={"config_list": config_list},
+    system_message=f'''
+    You are the Lead Copywriter. You MUST use the results from Agency_Researcher and Agency_Strategist in your response.
+    Craft compelling narratives and messages that align with {user_task} for {brand_task}.
+    Create differentiated content, impactful headlines, in-depth articles, ensuring that the brand's voice is consistent and compelling.
+    Work in tandem with the Agency Manager to ensure alignment with the user_proxy.
+    See {termination_notice}.
+    ''',
+    function_map={
+        "write_content": write_content,
+    }
+)
+
+writing_assistant = AssistantAgent(
+    name="writing_assistant",
+    llm_config=llm_config_content_assistant,
+    system_message=f'''
+    You are a writing assistant. You MUST use research function to collect the latest information about a given topic, 
+    and then use write_content function to write a pristine prose;
+    See {termination_notice}.
+    ''',
+    function_map={
+        "research": research,
+    }
+)
+
+agency_marketer = AssistantAgent(
+    name="Agency_Marketer",
+    llm_config={"config_list": config_list},
+    system_message=f'''
+    You are the Lead Marketer. 
+    Take the strategy and insights derived from research and transform them into compelling marketable ideas that resonate with the target audience.
+    Using the strategic direction from {user_task}, craft innovative marketing campaigns, promotions, and initiatives that effectively communicate the brand's value proposition.
+    Your expertise will bridge the gap between strategy and execution, ensuring that the brand's message is not only clear but also captivating. It's essential that your ideas are both impactful and aligned with the brand's overall vision.
+    Collaborate with other teams to ensure a cohesive approach, and always strive to push the boundaries of creativity to set our client's brand apart in the market.
+    Work in tandem with the Agency Manager to ensure alignment with the user_proxy.
+    See {termination_notice}.
+    '''
+)
+
+agency_mediaplanner = AssistantAgent(
+    name="Agency_Media_Planner",
+    llm_config={"config_list": config_list},
+    system_message=f'''
+    You are the Lead Media Planner. 
+    Identify the best mix of media channels to deliver an advertising message to a clients' target audience.
+    You MUST use the research function to query about contemporary and effective channels and tactics.
+    Using insights from {user_task}, strategize the most effective way to get a brand's message to its audience, whether it be through traditional media, digital platforms, or a combination of both.
+    Work in tandem with the Agency Manager to ensure alignment with the user_proxy.
+    See {termination_notice}.
+    ''',
+    function_map={
+        "research": research,
+    }
+)
+
+agency_director = AssistantAgent(
     name="Agency_Director",
     llm_config={"config_list": config_list},
     system_message=f'''
-    You are the Creative Director at SCTY. Be concise and avoid pleasantries. Refrain from any conversations that don't serve the goal of the user, ie. thank you.
-    Your primary role is to guide the creative vision of the project, ensuring that all ideas are not only unique and compelling but also meet the highest standards of excellence and desirability.
-    Drawing from the insights of {user_task}, oversee the creative process, inspire innovation, and set the bar for what's possible. Challenge the team to think outside the box and push the boundaries of creativity.
-    Review all creative outputs, provide constructive feedback, and ensure that every piece aligns with the brand's identity and resonates with the target audience. 
-    Your expertise is pivotal in ensuring that our work stands out in the market and leaves a lasting impact.
-    Collaborate closely with all teams, fostering a culture of excellence, and ensuring that our creative solutions are both groundbreaking and aligned with the project's objectives.
-    Engage with the Agency Strategist and Agency Marketer to ensure that the creative outputs align with the strategic direction and marketable ideas.
+    You are the Creative Director at SCTY. 
+    Critique and interrogate all work produced to ensure that all ideas are not only unique and compelling but also meet the highest standards of excellence and desirability.
+    Challenge the team to think outside the box and push the boundaries of creativity.
+    Work in tandem with the Agency Manager to ensure alignment with the user_proxy.
+    See {termination_notice}.
     '''
 )
 
-""" agency_accountmanager = autogen.AssistantAgent(
-    name="Agency_Account_Manager",
-    llm_config={"config_list": config_list},
-    system_message=f'''
-    You are the Account Manager.
-    Your primary responsibility is to nurture the relationship between the agency and the client, ensuring clear communication and understanding of the client's needs and feedback.
-    Act as a bridge between the client and the agency, facilitating collaboration, and ensuring that the project aligns with the client's expectations.
-    Regularly update the client on the project's status and ensure that their feedback is incorporated effectively.
-    Collaborate with the Agency Manager to ensure that timelines are met and deliverables are of the highest quality.
-    '''
-)
-
-agency_designer = autogen.AssistantAgent(
-    name="Agency_Designer",
-    llm_config={"config_list": config_list},
-    system_message=f'''
-    You are the Lead Designer.
-    Your primary responsibility is to transform strategic and marketing ideas into compelling visual narratives.
-    Drawing from the direction given in {user_task}, craft designs, layouts, and visual assets that align with the brand's identity and resonate with its target audience.
-    Work closely with the Creative Director and Agency Marketer to ensure that your designs align with the creative vision and marketing objectives.
-    Your expertise will ensure that our client's brand is visually captivating and stands out in the market.
-    '''
-) """
-
-user_proxy = autogen.UserProxyAgent(
-    name="User_proxy",
-    is_termination_msg=lambda x: x.get("content", "").strip().upper() == "TERMINATE",
-    human_input_mode="TERMINATE",
-    system_message="Provide the necessary research input. Save all output to markdown file",
-    code_execution_config={},
+user_proxy = UserProxyAgent(
+    name="user_proxy",
+    is_termination_msg=lambda msg: "TERMINATE" in msg["content"] if msg["content"] else False,
+    human_input_mode="ALWAYS",
+    max_consecutive_auto_reply=1,
+    code_execution_config={"work_dir": "/logs"},
+    system_message="Use the research and write_content functions as needed to answer the user's questions.",
     function_map={
         "research": research,
         "write_content": write_content,
+        "save_markdown": save_markdown,
     }
 )
 
@@ -410,10 +391,16 @@ user_proxy.register_function(
     }
 )
 
-groupchat = autogen.GroupChat(agents=[
-    user_proxy, agency_manager, agency_researcher, agency_strategist, agency_writer, writing_assistant, agency_marketer, agency_mediaplanner, agency_director], messages=[], max_round=20)
+groupchat = GroupChat(agents=[
+    user_proxy, agency_manager, agency_researcher, agency_strategist, agency_writer, writing_assistant, agency_marketer, agency_mediaplanner, agency_director], 
+    messages=[], 
+    max_round=20
+)
 
-manager = autogen.GroupChatManager(groupchat=groupchat, llm_config={"config_list": config_list})
+manager = GroupChatManager(
+    groupchat=groupchat, 
+    llm_config={"config_list": config_list}
+)
 
 user_proxy.initiate_chat(
     manager,
