+++ test/auto.py
